+++ lib/rapporteur/checks/active_record_check.rb
@@ -4,7 +4,7 @@ module Rapporteur
       def self.call(checker)
         ActiveRecord::Base.connection.execute("SELECT current_time AS time").first.fetch('time')
       rescue
-        checker.add_error(:database_unavailable)
+        checker.add_error(:database, :unavailable)
       end
     end
   end